:truck: [#1611] -- re-organize API data exchange code into package
Data fetching/pushing is now mostly organized in modules per aspect
in a single directory. This is a preparation step before refactoring
how a form is saved in the admin.
