+++ src/openforms/js/components/admin/form_design/data/index.js
@@ -1,0 +1,12 @@
+/**
+ * Package for data-interaction with the API endpoints.
+ *
+ * Note that this is still quite tightly coupled with the state of the form creation
+ * component, but the amount of code was getting unwieldy and introducing problems
+ * because of partial state updates.
+ */
+
+export {saveLogicRules, savePriceRules} from './logic';
+export {loadPlugins, PluginLoadingError} from './plugins';
+export {updateOrCreateFormSteps} from './steps';
+export {createOrUpdateFormVariables} from './variables';
